// src/enums/DKBCategory.ts
export enum DKBCategory {
  Regular = "Regular",
  Tutorial = "Tutorial",
  Triple = "Triple",
  Quintuple = "Quintuple",

  Battle = "Battle",
  Challenge = "Challenge",
  Cranky = "Cranky",
  Rehearsal = "Rehearsal",
  ShiftySmash = "Shifty Smash",

  Bloomintone = "Bloomintone",
  Chiptone = "Chiptone",
  Fragmentone = "Fragmentone",
  Glamtone = "Glamtone",
  Growtone = "Growtone",
  Quiztone = "Quiztone",
  Seekertone = "Seekertone",
  Smashintone = "Smashintone"
}